third_party/blink/web_tests/virtual/controls-refresh/week-picker/week-picker-ax.html

<!DOCTYPE html>
<html>
<head>
  <script src="../../../resources/testharness.js"></script>
  <script src="../../../resources/testharnessreport.js"></script>
  <script src="../../../fast/forms/resources/picker-common.js"></script>
  <script src="../../../fast/forms/calendar-picker/resources/calendar-picker-common.js"></script>
</head>

<body>
<input type="week" id="week1" value="2000-W13">

<script>
async_test((t) => {
    var markDirtyCounter = 0;
    var focusCounter = 0;
    accessibilityController.setNotificationListener(function(axnode, type) {
        if (type == 'Focus') {
            console.log('Focused: ' + accessibilityController.focusedElement.name.replace(/,/g, ''));
            focusCounter++;
        } else if (type == 'MarkDirty') {
            console.log('Received MarkDirty');
            markDirtyCounter++;
        }

        if (focusCounter == 3 && markDirtyCounter == 4) {
            t.done();
        }
    });

    var week1 = document.getElementById('week1');
    function test1() {
        assert_equals(focusCounter, 2, "Expected 2 Focus events from opening picker");
        assert_equals(markDirtyCounter, 3, "Expected 3 MarkDirty events from opening picker");
       eventSender.keyDown('ArrowDown');
    }

    openPicker(week1, test1);
}, 'Tests if typing an arrow key dispatches |Focus| and |MarkDirty| a11y events.');

</script>
</body>
</html>
